helloworld-pinecone/requirements.txt
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
pinecone-client
python-dotenv
